requirements.txt
aiohttp==3.6.2
bs4==0.0.1
discord.py==1.6.0
google-auth==1.6.2
google-api-python-client==1.8.2
python-dotenv==0.13.0
lxml==4.6.3
pytz==2020.1
anytree==2.8.0
